pyproject.toml
[tool.poetry]
name = "машиностроение"
version = "0.1.0"
description = ""
authors = ["Mail_Dilshat <mail.nuriev@narxoz.kz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.14.0"
django-allauth = "^0.54.0"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
